Feat: Component ZoneComponent
