+++ src/apis/requirementAPI.js
@@ -6,10 +6,3 @@ export const sendRequirement = (data) =>
       "Content-Type": `application/json`,
     },
   });
-
-export const sendRequirementDiagram = (data) =>
-  api.post("/api/v1/request-api/userRequirement", data, {
-    headers: {
-      "Content-Type": `application/json; charset=UTF-8`,
-    },
-  });
